+++ R/cpp11.R
@@ -244,8 +244,8 @@ cpp_numeric_to_int64 <- function(x) {
   .Call(`_cheapr_cpp_numeric_to_int64`, x)
 }
 
-cpp_format_double_as_int64 <- function(x) {
-  .Call(`_cheapr_cpp_format_double_as_int64`, x)
+cpp_format_numeric_as_int64 <- function(x) {
+  .Call(`_cheapr_cpp_format_numeric_as_int64`, x)
 }
 
 cpp_which_ <- function(x, invert) {